requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.11.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.7.1
distro==1.8.0
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.58.2
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.3.4
packaging==23.2
pathspec==0.11.2
pkgconfig==1.5.5
platformdirs==4.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
pymongo==4.6.0
PyMySQL==1.1.0
pyOpenSSL==23.3.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.1.0
zope.interface==6.1
